+++ index.js
@@ -7,6 +7,8 @@ const { refine } = require('./src/refine');
 const { recurse } = require('./src/recurse');
 const { buildSearchTool, buildExactTool, buildScanTool, litectxCorpus } = require('./src/recurse-retrieval');
 const { remember } = require('./src/remember');
+const { judge } = require('./src/judge');
+const { calibrate, CALIBRATION_CASES, INJECTION_BATTERY, scoreCase, gradeRun, constantHonored } = require('./src/judge-calibration');
 const { assessComplexity, isCritical } = require('./src/complexity');
 const { SkillRegistry } = require('./src/skills');
 const { createStashSkill } = require('./src/stash');
@@ -41,6 +43,13 @@ module.exports = {
   buildScanTool,
   litectxCorpus,
   remember,
+  judge,
+  calibrate,
+  CALIBRATION_CASES,
+  INJECTION_BATTERY,
+  scoreCase,
+  gradeRun,
+  constantHonored,
   assessComplexity,
   isCritical,
   SkillRegistry,